tech: moved more async methods
