+++ src/backend/main.ts
@@ -1,12 +1,6 @@
 import { initImagesCache } from './images_cache'
 import { downloadAntiCheatData } from './anticheat/utils'
-import {
-  AppSettings,
-  GamepadInputEvent,
-  GameSettings,
-  Runner,
-  StatusPromise
-} from 'common/types'
+import { GamepadInputEvent, StatusPromise } from 'common/types'
 import * as path from 'path'
 import { join } from 'path'
 import {
@@ -15,7 +9,6 @@ import {
   globalShortcut,
   ipcMain,
   Menu,
-  powerSaveBlocker,
   protocol,
   screen,
   session
@@ -23,16 +16,9 @@ import {
 import { uuid } from 'short-uuid'
 import { autoUpdater } from 'electron-updater'
 import { platform } from 'os'
-import {
-  appendFileSync,
-  existsSync,
-  rmSync,
-  watch,
-  writeFileSync
-} from 'graceful-fs'
+import { existsSync, rmSync, watch } from 'graceful-fs'
 import Backend from 'i18next-fs-backend'
 import i18next from 'i18next'
-import { GameConfig } from './game_config'
 import { GlobalConfig } from './config'
 import { LegendaryUser } from 'backend/storeManagers/legendary/user'
 import { GOGUser } from './storeManagers/gog/user'
@@ -48,7 +34,6 @@ import {
   configStore,
   createNecessaryFolders,
   eventsToCloseMetaMaskPopupOn,
-  fixAsarPath,
   gamesConfigPath,
   icon,
   installed,
@@ -56,7 +41,6 @@ import {
   isCLINoGui,
   isSteamDeckGameMode,
   publicDir,
-  tsStore,
   userHome
 } from './constants'
 import { handleOtp, handleProtocol } from './protocol'
@@ -65,11 +49,9 @@ import {
   logError,
   logInfo,
   LogPrefix,
-  logsDisabled,
   logWarning
 } from './logger/logger'
 import { gameInfoStore } from 'backend/storeManagers/legendary/electronStores'
-import shlex from 'shlex'
 import { initQueue } from './downloadmanager/downloadqueue'
 import {
   initOnlineMonitor,
@@ -77,7 +59,6 @@ import {
   runOnceWhenOnline
 } from './online_monitor'
 import { notify, showDialogBoxModalAuto } from './dialog/dialog'
-import { addRecentGame } from './recent_games/recent_games'
 import { initTrayIcon } from './tray_icon/tray_icon'
 import {
   createMainWindow,
@@ -86,25 +67,19 @@ import {
 } from './main_window'
 
 import * as HyperPlayLibraryManager from 'backend/storeManagers/hyperplay/library'
-import {
-  syncQueuedPlaytimeGOG,
-  updateGOGPlaytime
-} from 'backend/storeManagers/gog/games'
+import { syncQueuedPlaytimeGOG } from 'backend/storeManagers/gog/games'
 import { playtimeSyncQueue } from './storeManagers/gog/electronStores'
 import * as LegendaryLibraryManager from 'backend/storeManagers/legendary/library'
-import { getFlag, initLDClient } from './flags/flags'
+import { initLDClient } from './flags/flags'
 import { gameManagerMap, initStoreManagers } from './storeManagers'
 
 import * as Sentry from '@sentry/electron'
-import { DEV_PORTAL_URL, devSentryDsn, prodSentryDsn } from 'common/constants'
+import { devSentryDsn, prodSentryDsn } from 'common/constants'
 import { getHpOverlay, initOverlay } from './overlay'
 
 import { initExtension } from './extension/importer'
 import { hpApi } from './utils/hyperplay_api'
-import {
-  initializeCompatibilityLayer,
-  checkWineBeforeLaunch
-} from './utils/compatibility_layers'
+import { initializeCompatibilityLayer } from './utils/compatibility_layers'
 
 /*
  * INSERT OTHER IPC HANDLERS HERE
@@ -132,11 +107,6 @@ import { PROVIDERS } from 'common/types/proxy-types'
 import './ipcHandlers/index'
 
 import 'backend/utils/auto_launch'
-
-import { hrtime } from 'process'
-import { checkG7ConnectionStatus, postPlaySessionTime } from './utils/quests'
-
-import { gameIsEpicForwarderOnHyperPlay } from './utils/shouldOpenOverlay'
 
 import { createInjectedProviderWindow } from './injected_provider_window'
 
@@ -180,11 +150,6 @@ if (metricsAreEnabled()) {
   initSentry()
 }
 
-import { logFileLocation as getLogFileLocation } from './storeManagers/storeManagerCommon/games'
-import getPartitionCookies from './utils/get_partition_cookies'
-
-import { formatSystemInfo, getSystemInfo } from './utils/systeminfo'
-
 if (!app.isPackaged || process.env.DEBUG_HYPERPLAY === 'true') {
   app.commandLine?.appendSwitch('remote-debugging-port', '9222')
 
@@ -201,64 +166,6 @@ let ignoreExitToTray = false
 ipcMain.on('ignoreExitToTray', () => {
   ignoreExitToTray = true
 })
-
-async function completeHyperPlayQuest() {
-  const completeHpSummonQuestIsActive = getFlag(
-    'complete-hp-summon-quest',
-    false
-  )
-
-  if (!completeHpSummonQuestIsActive) {
-    return
-  }
-  logInfo('Completing HyperPlay Quest', LogPrefix.HyperPlay)
-  try {
-    const isConnected = await checkG7ConnectionStatus()
-    if (!isConnected) {
-      logInfo(
-        'HyperPlay account is not connected to Game7 Account',
-        LogPrefix.HyperPlay
-      )
-    }
-
-    const cookieString = await getPartitionCookies({
-      partition: 'persist:auth',
-      url: DEV_PORTAL_URL
-    })
-
-    const response = await fetch(`${DEV_PORTAL_URL}/api/hyperplay-quest`, {
-      method: 'POST',
-      headers: {
-        Cookie: cookieString
-      }
-    })
-
-    if (!response.ok) {
-      const error = await response.json()
-      logError(
-        `Failed to complete summon task: ${
-          error?.message ?? response.statusText
-        }`,
-        LogPrefix.HyperPlay
-      )
-      trackEvent({
-        event: 'HyperPlay Summon Quest Failed'
-      })
-      return
-    }
-
-    trackEvent({
-      event: 'HyperPlay Summon Quest Succeeded'
-    })
-  } catch (err) {
-    logError(`Error completing Summon quest ${err}`, LogPrefix.HyperPlay)
-    trackEvent({
-      event: 'HyperPlay Summon Quest Failed'
-    })
-  }
-
-  logInfo(`Completed HyperPlay Summon task`, LogPrefix.Backend)
-}
 
 async function initializeWindow(): Promise<BrowserWindow> {
   createNecessaryFolders()
@@ -692,60 +599,6 @@ app.on('open-url', (event, url) => {
 
 /// IPC handlers begin here.
 
-ipcMain.handle('getLocalPeloadPath', async () => {
-  return fixAsarPath(join(publicDir, 'webviewPreload.js'))
-})
-
-ipcMain.handle('requestSettings', async (event, appName) => {
-  // To the changes how we handle env and wrappers
-  // otherOptions is deprectaed and needs to be mapped
-  // to new approach.
-  // Can be removed if otherOptions is removed aswell
-  const mapOtherSettings = (config: AppSettings | GameSettings) => {
-    if (config.otherOptions) {
-      if (config.enviromentOptions.length <= 0) {
-        config.otherOptions
-          .split(' ')
-          .filter((val) => val.indexOf('=') !== -1)
-          .forEach((envKeyAndVar) => {
-            const keyAndValueSplit = envKeyAndVar.split('=')
-            const key = keyAndValueSplit.shift()!
-            const value = keyAndValueSplit.join('=')
-            config.enviromentOptions.push({ key, value })
-          })
-      }
-
-      if (config.wrapperOptions.length <= 0) {
-        const args: string[] = []
-        config.otherOptions
-          .split(' ')
-          .filter((val) => val.indexOf('=') === -1)
-          .forEach((val, index) => {
-            if (index === 0) {
-              config.wrapperOptions.push({ exe: val, args: '' })
-            } else {
-              args.push(val)
-            }
-          })
-
-        if (config.wrapperOptions.at(0)) {
-          config.wrapperOptions.at(0)!.args = shlex.join(args)
-        }
-      }
-
-      delete config.otherOptions
-    }
-    return config
-  }
-
-  if (appName === 'default') {
-    return mapOtherSettings(GlobalConfig.get().getSettings())
-  }
-
-  const config = await GameConfig.get(appName).getSettings()
-  return mapOtherSettings(config)
-})
-
 // Watch the installed games file and trigger a refresh on the installed games if something changes
 if (existsSync(installed)) {
   let watchTimeout: NodeJS.Timeout | undefined
@@ -758,309 +611,6 @@ if (existsSync(installed)) {
     watchTimeout = setTimeout(LegendaryLibraryManager.refreshInstalled, 500)
   })
 }
-
-let powerDisplayId: number | null
-let gamePlaySessionStartTimes: Record<string, bigint> = {}
-
-function startNewPlaySession(appName: string) {
-  const prevStartTime = gamePlaySessionStartTimes[appName]
-  gamePlaySessionStartTimes = {}
-  // Uses hrtime for monotonic timer not subject to clock drift or sync errors
-  const startPlayingTimeMonotonic = hrtime.bigint()
-  gamePlaySessionStartTimes[appName] = startPlayingTimeMonotonic
-  return prevStartTime
-}
-
-async function postPlaySession(
-  appName: string,
-  runner: Runner,
-  sessionPlaytimeInMs: bigint
-) {
-  const game = gameManagerMap[runner].getGameInfo(appName)
-  const { hyperPlayListing } = await gameIsEpicForwarderOnHyperPlay(game)
-  await postPlaySessionTime(
-    hyperPlayListing?.project_id || appName,
-    // round up to prevent session time loss
-    parseInt(((sessionPlaytimeInMs + BigInt(1000)) / BigInt(1000)).toString())
-  )
-}
-
-function syncPlaySession(appName: string) {
-  if (!Object.hasOwn(gamePlaySessionStartTimes, appName)) {
-    return BigInt(0)
-  }
-
-  // reset the time counter and start new session slightly before ending current session to prevent time loss
-  const startPlayingTimeMonotonic = startNewPlaySession(appName)
-  const stopPlayingTimeMonotonic = hrtime.bigint()
-  const sessionPlaytimeInMs =
-    (stopPlayingTimeMonotonic - startPlayingTimeMonotonic) / BigInt(1000000)
-
-  // update local json with time played
-  const sessionPlaytimeInMinutes =
-    sessionPlaytimeInMs / BigInt(1000) / BigInt(60)
-
-  const totalPlaytime =
-    sessionPlaytimeInMinutes + BigInt(tsStore.get(`${appName}.totalPlayed`, 0))
-  tsStore.set(`${appName}.totalPlayed`, Number(totalPlaytime))
-
-  return sessionPlaytimeInMs
-}
-
-ipcMain.handle(
-  'syncPlaySession',
-  async (e, appName: string, runner: Runner) => {
-    const sessionPlaytimeInMs = syncPlaySession(appName)
-    await postPlaySession(appName, runner, sessionPlaytimeInMs)
-  }
-)
-
-// get pid/tid on launch and inject
-ipcMain.handle(
-  'launch',
-  async (event, { appName, launchArguments, runner }): StatusPromise => {
-    const game = gameManagerMap[runner].getGameInfo(appName)
-    const gameSettings = await gameManagerMap[runner].getSettings(appName)
-    const { autoSyncSaves, savesPath, gogSaves = [] } = gameSettings
-
-    const { title, app_name, browserUrl, install } = game
-
-    const { minimizeOnGameLaunch } = GlobalConfig.get().getSettings()
-
-    const startPlayingDate = new Date()
-    startNewPlaySession(appName)
-
-    if (!tsStore.has(game.app_name)) {
-      tsStore.set(
-        `${game.app_name}.firstPlayed`,
-        startPlayingDate.toISOString()
-      )
-    }
-
-    logInfo(`Launching ${title} (${game.app_name})`, LogPrefix.Backend)
-    trackEvent({
-      event: 'Game Launched',
-      properties: {
-        game_name: app_name,
-        isBrowserGame: browserUrl !== undefined,
-        game_title: title,
-        store_name: getStoreName(runner),
-        browserUrl: browserUrl ?? undefined,
-        platform: getPlatformName(install.platform!),
-        platform_arch: install.platform!
-      }
-    })
-
-    // purposefully not awaiting this
-    completeHyperPlayQuest()
-
-    if (autoSyncSaves && isOnline()) {
-      sendFrontendMessage('gameStatusUpdate', {
-        appName,
-        runner,
-        status: 'syncing-saves'
-      })
-      logInfo(`Downloading saves for ${title}`, LogPrefix.Backend)
-      try {
-        await gameManagerMap[runner].syncSaves(
-          appName,
-          '--skip-upload',
-          savesPath,
-          gogSaves
-        )
-        logInfo(`Saves for ${title} downloaded`, LogPrefix.Backend)
-      } catch (error) {
-        logError(
-          `Error while downloading saves for ${title}. ${error}`,
-          LogPrefix.Backend
-        )
-      }
-    }
-
-    sendFrontendMessage('gameStatusUpdate', {
-      appName,
-      runner,
-      status: 'playing'
-    })
-
-    const mainWindow = getMainWindow()
-    if (minimizeOnGameLaunch) {
-      mainWindow?.hide()
-    }
-
-    // Prevent display from sleep
-    if (!powerDisplayId) {
-      logInfo('Preventing display from sleep', LogPrefix.Backend)
-      powerDisplayId = powerSaveBlocker.start('prevent-display-sleep')
-    }
-
-    const logFileLocation = getLogFileLocation(appName)
-
-    const systemInfo = await getSystemInfo()
-      .then(formatSystemInfo)
-      .catch((error) => {
-        logError(
-          ['Failed to fetch system information', error],
-          LogPrefix.Backend
-        )
-        return 'Error, check general log'
-      })
-    writeFileSync(logFileLocation, 'System Info:\n' + `${systemInfo}\n` + '\n')
-
-    const gameSettingsString = JSON.stringify(gameSettings, null, '\t')
-    appendFileSync(
-      logFileLocation,
-      'System Info:\n' +
-        `${systemInfo}\n` +
-        '\n' +
-        `Game Settings: ${gameSettingsString}\n` +
-        '\n' +
-        `Game launched at: ${startPlayingDate}\n` +
-        '\n'
-    )
-
-    if (logsDisabled) {
-      appendFileSync(
-        logFileLocation,
-        'IMPORTANT: Logs are disabled. Enable logs before reporting an issue.'
-      )
-    }
-
-    const isNative = gameManagerMap[runner].isNative(appName)
-
-    // check if isNative, if not, check if wine is valid
-    if (!isNative) {
-      const isWineOkToLaunch = await checkWineBeforeLaunch(
-        appName,
-        gameSettings,
-        logFileLocation
-      )
-
-      if (!isWineOkToLaunch) {
-        logError(
-          `Was not possible to launch using ${gameSettings.wineVersion.name}`,
-          LogPrefix.Backend
-        )
-
-        sendFrontendMessage('gameStatusUpdate', {
-          appName,
-          runner,
-          status: 'done'
-        })
-
-        return { status: 'error' }
-      }
-    }
-
-    sendFrontendMessage('gameStatusUpdate', {
-      appName,
-      runner,
-      status: 'playing'
-    })
-
-    const command = gameManagerMap[runner].launch(appName, launchArguments)
-
-    const launchResult = await command.catch((exception) => {
-      logError(exception, LogPrefix.Backend)
-      appendFileSync(
-        logFileLocation,
-        `An exception occurred when launching the game:\n${exception.stack}`
-      )
-      return false
-    })
-
-    // Stop display sleep blocker
-    if (powerDisplayId !== null) {
-      logInfo('Stopping Display Power Saver Blocker', LogPrefix.Backend)
-      powerSaveBlocker.stop(powerDisplayId)
-    }
-
-    // Update playtime and last played date
-    const finishedPlayingDate = new Date()
-    tsStore.set(`${appName}.lastPlayed`, finishedPlayingDate.toISOString())
-
-    if (runner === 'gog') {
-      await updateGOGPlaytime(appName, startPlayingDate, finishedPlayingDate)
-    }
-
-    await addRecentGame(game)
-
-    if (autoSyncSaves && isOnline()) {
-      /**
-       * @dev It sets to done, so the GlobalState knows that the game session stopped.
-       * Then it changes the status to syncing-saves. Then It sets to done again.
-       * Otherwise it would count the Syncing Saves time (which can be long depending on the game) as playing time as well.
-       * done is not only the state for stopping playing but for finishing any other process that came before.
-       */
-      sendFrontendMessage('gameStatusUpdate', {
-        appName,
-        runner,
-        status: 'done'
-      })
-
-      sendFrontendMessage('gameStatusUpdate', {
-        appName,
-        runner,
-        status: 'syncing-saves'
-      })
-
-      sendFrontendMessage('gameStatusUpdate', {
-        appName,
-        runner,
-        status: 'done'
-      })
-
-      logInfo(`Uploading saves for ${title}`, LogPrefix.Backend)
-      try {
-        await gameManagerMap[runner].syncSaves(
-          appName,
-          '--skip-download',
-          savesPath,
-          gogSaves
-        )
-        logInfo(`Saves uploaded for ${title}`, LogPrefix.Backend)
-      } catch (error) {
-        logError(
-          `Error uploading saves for ${title}. Error: ${error}`,
-          LogPrefix.Backend
-        )
-      }
-    }
-
-    sendFrontendMessage('gameStatusUpdate', {
-      appName,
-      runner,
-      status: 'done'
-    })
-
-    // Playtime of this session in milliseconds. Uses hrtime for monotonic timer not subject to clock drift or sync errors
-    const sessionPlaytimeInMs = syncPlaySession(appName)
-    postPlaySession(appName, runner, sessionPlaytimeInMs)
-
-    trackEvent({
-      event: 'Game Closed',
-      properties: {
-        game_name: app_name,
-        isBrowserGame: browserUrl !== undefined,
-        game_title: title,
-        store_name: getStoreName(runner),
-        browserUrl: browserUrl ?? undefined,
-        platform: getPlatformName(install.platform!),
-        playTimeInMs: Number(sessionPlaytimeInMs),
-        platform_arch: install.platform!
-      }
-    })
-
-    // Exit if we've been launched without UI
-    if (isCLINoGui) {
-      app.exit()
-    } else {
-      mainWindow?.show()
-    }
-
-    return { status: launchResult ? 'done' : 'error' }
-  }
-)
 
 ipcMain.handle(
   'uninstall',
